+++ zebra-rpc/src/server.rs
@@ -7,15 +7,14 @@
 //! See the full list of
 //! [Differences between JSON-RPC 1.0 and 2.0.](https://www.simple-is-better.org/rpc/#differences-between-1-0-and-2-0)
 
-use std::{panic, sync::Arc};
+use std::{fmt, panic, sync::Arc};
 
 use jsonrpc_core::{Compatibility, MetaIoHandler};
-use jsonrpc_http_server::ServerBuilder;
+use jsonrpc_http_server::{CloseHandle, ServerBuilder};
 use tokio::task::JoinHandle;
 use tower::{buffer::Buffer, Service};
 
-use tracing::*;
-use tracing_futures::Instrument;
+use tracing::{Instrument, *};
 
 use zebra_chain::{
     block::{self, Block},
@@ -31,7 +30,7 @@ use crate::{
 };
 
 #[cfg(feature = "getblocktemplate-rpcs")]
-use crate::methods::{GetBlockTemplateRpc, GetBlockTemplateRpcImpl};
+use crate::methods::{get_block_template_rpcs, GetBlockTemplateRpc, GetBlockTemplateRpcImpl};
 
 pub mod compatibility;
 mod tracing_middleware;
@@ -40,13 +39,45 @@ mod tracing_middleware;
 mod tests;
 
 /// Zebra RPC Server
-#[derive(Clone, Debug)]
-pub struct RpcServer;
+#[derive(Clone)]
+pub struct RpcServer {
+    config: Config,
+    network: Network,
+    app_version: String,
+    close_handle: CloseHandle,
+}
+
+impl fmt::Debug for RpcServer {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        f.debug_struct("RpcServer")
+            .field("config", &self.config)
+            .field("network", &self.network)
+            .field("app_version", &self.app_version)
+            .field(
+                "close_handle",
+                // TODO: when it stabilises, use std::any::type_name_of_val(&self.close_handle)
+                &"CloseHandle",
+            )
+            .finish()
+    }
+}
 
 impl RpcServer {
-    /// Start a new RPC server endpoint
+    /// Start a new RPC server endpoint using the supplied configs and services.
+    /// `app_version` is a version string for the application, which is used in RPC responses.
+    ///
+    /// Returns [`JoinHandle`]s for the RPC server and `sendrawtransaction` queue tasks,
+    /// and a [`RpcServer`] handle, which can be used to shut down the RPC server task.
+    //
+    // TODO: put some of the configs or services in their own struct?
+    #[allow(clippy::too_many_arguments)]
     pub fn spawn<Version, Mempool, State, Tip, ChainVerifier>(
         config: Config,
+        #[cfg(feature = "getblocktemplate-rpcs")]
+        mining_config: get_block_template_rpcs::config::Config,
+        #[cfg(not(feature = "getblocktemplate-rpcs"))]
+        #[allow(unused_variables)]
+        mining_config: (),
         app_version: Version,
         mempool: Buffer<Mempool, mempool::Request>,
         state: State,
@@ -54,9 +85,9 @@ impl RpcServer {
         chain_verifier: ChainVerifier,
         latest_chain_tip: Tip,
         network: Network,
-    ) -> (JoinHandle<()>, JoinHandle<()>)
+    ) -> (JoinHandle<()>, JoinHandle<()>, Option<Self>)
     where
-        Version: ToString + Clone,
+        Version: ToString + Clone + Send + 'static,
         Mempool: tower::Service<
                 mempool::Request,
                 Response = mempool::Response,
@@ -91,6 +122,8 @@ impl RpcServer {
             {
                 // Initialize the getblocktemplate rpc method handler
                 let get_block_template_rpc_impl = GetBlockTemplateRpcImpl::new(
+                    network,
+                    mining_config,
                     mempool.clone(),
                     state.clone(),
                     latest_chain_tip.clone(),
@@ -102,7 +135,7 @@ impl RpcServer {
 
             // Initialize the rpc methods with the zebra version
             let (rpc_impl, rpc_tx_queue_task_handle) = RpcImpl::new(
-                app_version,
+                app_version.clone(),
                 network,
                 config.debug_force_finished_sync,
                 mempool,
@@ -119,18 +152,14 @@ impl RpcServer {
             }
 
             // The server is a blocking task, which blocks on executor shutdown.
-            // So we need to create and spawn it on a std::thread, inside a tokio blocking task.
-            // (Otherwise tokio panics when we shut down the RPC server.)
+            // So we need to start it in a std::thread.
+            // (Otherwise tokio panics on RPC port conflict, which shuts down the RPC server.)
             let span = Span::current();
-            let server = move || {
+            let start_server = move || {
                 span.in_scope(|| {
                     // Use a different tokio executor from the rest of Zebra,
                     // so that large RPCs and any task handling bugs don't impact Zebra.
-                    //
-                    // TODO:
-                    // - return server.close_handle(), which can shut down the RPC server,
-                    //   and add it to the server tests
-                    let server = ServerBuilder::new(io)
+                    let server_instance = ServerBuilder::new(io)
                         .threads(parallel_cpu_threads)
                         // TODO: disable this security check if we see errors from lightwalletd
                         //.allowed_hosts(DomainsValidation::Disabled)
@@ -138,32 +167,117 @@ impl RpcServer {
                         .start_http(&listen_addr)
                         .expect("Unable to start RPC server");
 
-                    info!("Opened RPC endpoint at {}", server.address());
+                    info!("Opened RPC endpoint at {}", server_instance.address());
 
-                    server.wait();
+                    let close_handle = server_instance.close_handle();
 
-                    info!("Stopping RPC endpoint");
+                    let rpc_server_handle = RpcServer {
+                        config,
+                        network,
+                        app_version: app_version.to_string(),
+                        close_handle,
+                    };
+
+                    (server_instance, rpc_server_handle)
                 })
             };
 
-            (
-                tokio::task::spawn_blocking(|| {
-                    let thread_handle = std::thread::spawn(server);
+            // Propagate panics from the std::thread
+            let (server_instance, rpc_server_handle) = match std::thread::spawn(start_server).join()
+            {
+                Ok(rpc_server) => rpc_server,
+                Err(panic_object) => panic::resume_unwind(panic_object),
+            };
 
-                    // Propagate panics from the inner std::thread to the outer tokio blocking task
-                    match thread_handle.join() {
-                        Ok(()) => (),
-                        Err(panic_object) => panic::resume_unwind(panic_object),
-                    }
-                }),
+            // The server is a blocking task, which blocks on executor shutdown.
+            // So we need to wait on it on a std::thread, inside a tokio blocking task.
+            // (Otherwise tokio panics when we shut down the RPC server.)
+            let span = Span::current();
+            let wait_on_server = move || {
+                span.in_scope(|| {
+                    server_instance.wait();
+
+                    info!("Stopped RPC endpoint");
+                })
+            };
+
+            let span = Span::current();
+            let rpc_server_task_handle = tokio::task::spawn_blocking(move || {
+                let thread_handle = std::thread::spawn(wait_on_server);
+
+                // Propagate panics from the inner std::thread to the outer tokio blocking task
+                span.in_scope(|| match thread_handle.join() {
+                    Ok(()) => (),
+                    Err(panic_object) => panic::resume_unwind(panic_object),
+                })
+            });
+
+            (
+                rpc_server_task_handle,
                 rpc_tx_queue_task_handle,
+                Some(rpc_server_handle),
             )
         } else {
             // There is no RPC port, so the RPC tasks do nothing.
             (
                 tokio::task::spawn(futures::future::pending().in_current_span()),
                 tokio::task::spawn(futures::future::pending().in_current_span()),
+                None,
             )
         }
+    }
+
+    /// Shut down this RPC server, blocking the current thread.
+    ///
+    /// This method can be called from within a tokio executor without panicking.
+    /// But it is blocking, so `shutdown()` should be used instead.
+    pub fn shutdown_blocking(&self) {
+        Self::shutdown_blocking_inner(self.close_handle.clone())
+    }
+
+    /// Shut down this RPC server asynchronously.
+    /// Returns a task that completes when the server is shut down.
+    pub fn shutdown(&self) -> JoinHandle<()> {
+        let close_handle = self.close_handle.clone();
+
+        let span = Span::current();
+        tokio::task::spawn_blocking(move || {
+            span.in_scope(|| Self::shutdown_blocking_inner(close_handle))
+        })
+    }
+
+    /// Shuts down this RPC server using its `close_handle`.
+    ///
+    /// See `shutdown_blocking()` for details.
+    fn shutdown_blocking_inner(close_handle: CloseHandle) {
+        // The server is a blocking task, so it can't run inside a tokio thread.
+        // See the note at wait_on_server.
+        let span = Span::current();
+        let wait_on_shutdown = move || {
+            span.in_scope(|| {
+                info!("Stopping RPC server");
+                close_handle.clone().close();
+                info!("Stopped RPC server");
+            })
+        };
+
+        let span = Span::current();
+        let thread_handle = std::thread::spawn(wait_on_shutdown);
+
+        // Propagate panics from the inner std::thread to the outer tokio blocking task
+        span.in_scope(|| match thread_handle.join() {
+            Ok(()) => (),
+            Err(panic_object) => panic::resume_unwind(panic_object),
+        })
+    }
+}
+
+impl Drop for RpcServer {
+    fn drop(&mut self) {
+        // Block on shutting down, propagating panics.
+        // This can take around 150 seconds.
+        //
+        // Without this shutdown, Zebra's RPC unit tests sometimes crashed with memory errors.
+        self.shutdown_blocking();
     }
 }